obs-lab/pe-lab/runner/requirements.txt
openai>=1.0.0
kafka-python